refactor: change button group as legacy
